fix: highlighter without memoizing
