+++ src/chart_types/heatmap/renderer/dom/highlighter_brush.tsx
@@ -22,6 +22,7 @@ import { GlobalChartState } from '../../../../state/chart_state';
 import { getInternalIsInitializedSelector, InitStatus } from '../../../../state/selectors/get_internal_is_intialized';
 import { computeChartDimensionsSelector } from '../../state/selectors/compute_chart_dimensions';
 import { geometries } from '../../state/selectors/geometries';
+import { getBrushedHighlightedShapesSelector } from '../../state/selectors/get_brushed_highlighted_shapes';
 import { getHighlightedAreaSelector } from '../../state/selectors/get_highlighted_area';
 import { DEFAULT_PROPS, HighlighterCellsComponent, HighlighterCellsProps } from './highlighter';
 
@@ -34,9 +35,8 @@ const brushMapStateToProps = (state: GlobalChartState): HighlighterCellsProps =>
 
   const geoms = geometries(state);
   const canvasDimension = computeChartDimensionsSelector(state);
-  // FIXME stateful picked area from the internal state
-  // @ts-ignore
-  let dragShape = state.internalChartState.getPickedArea(state);
+
+  let dragShape = getBrushedHighlightedShapesSelector(state);
   const highlightedArea = getHighlightedAreaSelector(state);
   if (highlightedArea) {
     dragShape = highlightedArea;
